+++ exts/wandelbots.omni/wandelbots/omni/gui/auth.py
@@ -1,10 +1,9 @@
-import asyncio
 import omni.ui as ui
 import webbrowser
+import asyncio
 import carb
+from wandelbots.omni.utils.auth import store_auth_token, poll_token_endpoint, get_device_code_info, get_auth_config
 from nova.auth.authorization import Auth0DeviceAuthorization
-from wandelbots.omni.utils.auth import Auth0Model
-from wandelbots.omni.environment import load_env
 
 class Auth0UIBuilder:
     def __init__(self):
@@ -16,9 +15,25 @@ class Auth0UIBuilder:
         self._auth_subline_lbl = "This code must be identical to the code displayed during the instance confirmation:"
         self._success_headline_lbl = "Authentication successful. Happy simulating!"
         self._error_headline_lbl = "Authentication not possible. Please try again later."
+        self._auth_config = get_auth_config()
+        self._auth_controller = Auth0DeviceAuthorization(self._auth_config)
 
     def _on_dismissed_window(self, is_visible):
         self._dismissed = not is_visible
+
+    def _handle_button_click(self, url):
+        """Helper function to handle button click"""
+       
+        try:
+            loop = asyncio.get_event_loop()
+        except RuntimeError:
+            # Create a new event loop if one doesn't exist
+            loop = asyncio.new_event_loop()
+            asyncio.set_event_loop(loop)
+        
+        webbrowser.open(url)
+        # Create task in the event loop
+        loop.create_task(self.check_auth_status())
 
     def create_initial_view(self, device_code_info, open_verification_url):
         with self._auth_window.frame:
@@ -44,7 +59,7 @@ class Auth0UIBuilder:
                         self._auth_button_lbl, width=300, height=50
                     )
                     open_url_button.set_clicked_fn(
-                        lambda: webbrowser.open(open_verification_url)
+                        lambda: self._handle_button_click(open_verification_url)
                     )
                     ui.Spacer()
                 ui.Label(
@@ -94,54 +109,40 @@ class Auth0UIBuilder:
 
                 return close_button
 
-    def create_auth_ui(self, device_code_info, auth0_device_auth):
+    def create_auth_ui(self):
         self._auth_window = ui.Window(
             self._window_title, width=400, height=240, visibility_changed_fn=self._on_dismissed_window
         )
+        device_code_info = get_device_code_info(self._auth_controller)
         url = f"{device_code_info.verification_uri}?user_code={device_code_info.user_code}"
         self.create_initial_view(device_code_info, url)
 
-        async def check_auth_status():
-            try:
-                token = await auth0_device_auth.poll_token_endpoint()
-                Auth0Model.store_token(token)
-                self._auth_window.frame.clear()
-                if self._dismissed: 
-                    return
-                close_button = self.create_success_view()
-                close_button.set_clicked_fn(
-                    lambda: setattr(self._auth_window, "visible", False)
-                )
-            except Exception:
-                Auth0Model.store_token(None)
-                self._auth_window.frame.clear()
-                if self._dismissed: 
-                    return
-                close_button = self.create_error_view()
-                close_button.set_clicked_fn(
-                    lambda: setattr(self._auth_window, "visible", False)
-                )
-
-        # Schedule the check_auth_status coroutine
-        asyncio.create_task(check_auth_status())
+    async def check_auth_status(self):
+        """Async function to check authentication status and update UI"""
+        try:
+            carb.log_info("Starting token polling...")
+            token = await poll_token_endpoint(self._auth_controller)
+            store_auth_token(token)
+            self._auth_window.frame.clear()
+            if self._dismissed: 
+                return
+            close_button = self.create_success_view()
+            close_button.set_clicked_fn(
+                lambda: setattr(self._auth_window, "visible", False)
+            )
+        except Exception as e:
+            carb.log_error(f"Error during authentication: {e}")
+            self._auth_window.frame.clear()
+            if self._dismissed: 
+                return
+            close_button = self.create_error_view()
+            close_button.set_clicked_fn(
+                lambda: setattr(self._auth_window, "visible", False)
+            )
 
     def display_auth_window(self):
-        config = load_env()
-        auth0_domain = config("AUTH0_DOMAIN")
-        auth0_client_id = config("AUTH0_CLIENT_ID")
-        auth0_audience = config("AUTH0_AUDIENCE")
-
-        auth0_device_auth = Auth0DeviceAuthorization(
-            auth0_domain, auth0_client_id, auth0_audience
-        )
-
-        try:
-            # Request a device code
-            device_code_info = auth0_device_auth.request_device_code()
-            self.create_auth_ui(device_code_info, auth0_device_auth)
-
-        except Exception as e:
-            carb.log_error(f"An error occurred: {e}")
+        self.create_auth_ui()
 
 
 
+