+++ exts/wandelbots.omni/wandelbots/omni/utils/api.py
@@ -1,10 +1,13 @@
 import wandelbots_api_client as wb
-from .auth import Auth0Model
 
-def get_api_client(host: str, secure=False) -> wb.ApiClient:
+def get_api_client(host: str, secure=False, token: str | None = None) -> wb.ApiClient:
     base_url = f"http{'s' if secure else ''}://{host}/api/v1"
     config = wb.Configuration(
         host=base_url,
-        access_token=Auth0Model.get_token()
+        access_token=token
     )
-    return wb.ApiClient(config)
+    return wb.ApiClient(
+        configuration=config,
+        header_name="X-Wandelbots-Client",
+        header_value="isaac-sim-extension"
+    )